Creating the application infrastructure and placeholder subcommand.
This is one large commit because of how interdependent these objects are. The crate::app module should contain all of the primitives necessary to create a subcommand. The architectural differences from the previous version primarily boil down to a simple re-organization but also a redesign on how errors and subcommand contexts are handled. In this version, the context is a single object instead of multiple being passed to each command. This affords us flexibility in adding new context fields if it ever becomes necessary. The Error struct has also been upgraded so that we can provide additional context to each other. Previously, vague OS level errors would simply be displayed and trying to add more context would have complicated each subcommand trying to add it.